clean up models introducing base  and using uuid.hex as id with optional prefix
