+++ purpleserver/core/purpleserver/core/admin.py
@@ -1,6 +1,7 @@
+import inspect
 from django.contrib import admin
+import purpleserver.core.models as models
 
-from purpleserver.core.models import MODELS
-
-for model in MODELS.values():
-    admin.site.register(model)
+for name, model in inspect.getmembers(models):
+    if inspect.isclass(model) and issubclass(model, models.Carrier) and name not in ['Carrier']:
+        admin.site.register(model)
