+++ purpleserver/core/purpleserver/core/models/entity.py
@@ -1,0 +1,26 @@
+from uuid import uuid4
+from django.db import models
+from django.conf import settings
+
+
+def uuid(prefix: str = None):
+    return f'{prefix or ""}{uuid4().hex}'
+
+
+class Entity(models.Model):
+    class Meta:
+        abstract = True
+
+    id = models.CharField(max_length=50, primary_key=True, default=uuid, editable=False)
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+
+    def __str__(self):
+        return self.id
+
+
+class OwnedEntity(Entity):
+    class Meta:
+        abstract = True
+
+    user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
